modified Problem B

+++ 2017RoundB/ProblemB/ProblemB.cpp
@@ -11,18 +11,18 @@ bool sort_X(vector<float> a, vector<float> b) {
 bool sort_Y(vector<float> a, vector<float> b) {
 	return a[1]<b[1];
 }
-float center(vector<vector<float>> &nums) {
+double center(vector<vector<float>> &nums) {
 	int n = nums.size();
 	//transfer to manhaton distance
 	for (int i = 0; i < n; i++) {
-		float x = nums[i][0], y = nums[i][1];
+		double x = nums[i][0], y = nums[i][1];
 		nums[i][0] = x + y;
 		nums[i][1] = x - y;
 	}
 	//calculate x
 	sort(nums.begin(), nums.end(), sort_X);
 	int start = 0, end = n - 1;
-	float left = nums[start][2], right = nums[end][2];
+	double left = nums[start][2], right = nums[end][2];
 	while (start<end) {
 		if (left<right) {
 			start++;
@@ -49,7 +49,7 @@ float center(vector<vector<float>> &nums) {
 		}
 	}
 	float Y = nums[start][1];
-	float res = 0;
+	double res = 0;
 	for (int i = 0; i < n; i++) {
 		float x = nums[i][0], y = nums[i][1], w = nums[i][2];
 		res += (abs(x - X) + abs(y - Y))*w;
